+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - Cluster specific agents tab.
 - Cluster specific events tab.
+- Cluster specific nodes tab.
 
 ### Changed
 - Improve loading of cluster overview page.